Removed needless separation; fixed state
The original version used a container/component architecture under the
impression that the *container* would manage state and the *component*
would display it. With the state in a Om-Next atom, there's no need for
the extra overhead.

The state in v1 was also flat, single-level. Here, the media-list state
is in one sub-map, and the media-player state is in another.